Separate out the sentence
